+++ src/buzz_api.zig
@@ -241,7 +241,7 @@ fn valueDump(value: Value, vm: *VM, seen: *std.AutoHashMap(*_obj.Obj, void), dep
                                 "{s}{s}{s}: {s}",
                                 .{
                                     if (field.static) "static" else "",
-                                    if (field.constant) "const" else "",
+                                    if (field.final) "final" else "",
                                     field.name,
                                     field_type_str.items,
                                 },